Colours for scale_fill_gradientn() in "plot_station_horizon_range"

Colour palette is attributed for a set of range values in metre and changes at 100 m corresponding to the default distance to download high resolution DEMs

colour_range

Format

An object of class list of length 2.

Examples

colour_range
#> $colour
#>  [1] "#E31A1C" "#FB9A99" "#33A02C" "#B2DF8A" "#1F78B4" "#A6CEE3" "#F2F0F7"
#>  [8] "#CBC9E2" "#9E9AC8" "#6A51A3"
#> 
#> $range
#>  [1]     0     5    10    30    50    70    90   100  1000  5000 20000
#> 

## Generated from
# c(rev(RColorBrewer::brewer.pal(6, "Paired")),
#       RColorBrewer::brewer.pal(4, "Purples"))