Compute siting class for a station measuring air temperature

Assess the class of a station from the type of surrounding land cover, slope, vegetation and shadowing as potential heat sources affecting sensors and using the requirements set by WMO and Met

compute_class_precipitation(
  stn = NULL,
  horizon = NULL,
  test_type = "MET",
  f_plot = TRUE
)

Arguments

stn

A SpatVector with station attributes from "get_latlon_frost"

horizon

A data.frame with horizon distribution from "compute_horizon_max"

test_type

A string defining a type of tests to apply: "WMO" or "MET"

f_plot

A boolean whether to plot figures

Value

A class

References

https://community.wmo.int/en/activity-areas/imop/siting-classification

Examples

# Get station metadata
stn <- get_metadata_frost(stationid = 18700, dx = 100, resx = 1)
#> [1] " "
#> [1] "-------------------------------------------"
#> [1] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [2] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [3] "station MET.NO: 18700 -- OSLO - BLINDERN -- WMO: 0-20000-0-01492"      
#> [4] "station MET.NO: 18700 -- OSLO - BLINDERN -- WIGOS: 0-20000-0-01492"    
#> [1] "-------------------------------------------"
#> [1] " "

# Compute maximum horizon
horizon_max <- compute_horizon_max(stn, step = 1, f_plot_polygon = FALSE)
#> [1] "Process: 18700 - 260966.8/6652718.0 - dtm - 100/1 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dtm_25833_d00100m_1.0m.tif"
#> [1] "Process: 18700 - 260966.8/6652718.0 - dom - 100/1 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dom_25833_d00100m_1.0m.tif"
#> [1] "Process: 18700 - 260966.8/6652718.0 - dtm - 20000/20 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dtm_25833_d20000m_20.0m.tif"
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory

# Compute class
compute_class_precipitation(stn,
                            horizon_max,
                            test_type = "WMO",
                            f_plot = TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'global': error in evaluating the argument 'x' in selecting a method for function 'terrain': object 'dem' not found