Maximum horizon height around a station

Compute maximum horizon height in degrees from a station location in UTM 33 and based on three DEMs: Digital Terrain Model, Digital Surface Model and 20km-Digital Terrain Model

compute_horizon_max(
  stn = NULL,
  step = 10,
  f_plot_polygon = FALSE,
  f_output_all = FALSE
)

Arguments

stn

A SpatVector with station attributes from "get_latlon_frost"

step

An interval in degrees at which the horizon will be computed, default is every 10 deg.

f_plot_polygon

A boolean setting boundary values to -20 deg to facilitate plotting as polygon

f_output_all

A boolean to export all horizons: dem, dsm, demkm & max

Value

A dataframe with `horizon_height` in degrees and `azimuth` angle in degrees at which the horizon is computed in degrees

References

https://grass.osgeo.org/grass83/manuals/r.horizon.html

Examples

# Load the station metadata including location and level
stn <- get_metadata_frost(stationid = 18700, dx = 100, resx = 1)
#> [1] " "
#> [1] "-------------------------------------------"
#> [1] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [2] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [3] "station MET.NO: 18700 -- OSLO - BLINDERN -- WMO: 0-20000-0-01492"      
#> [4] "station MET.NO: 18700 -- OSLO - BLINDERN -- WIGOS: 0-20000-0-01492"    
#> [1] "-------------------------------------------"
#> [1] " "

# Compute the maximum horizon
compute_horizon_max(stn)
#> [1] "Process: 18700 - 260966.8/6652718.0 - dtm - 100/1 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dtm_25833_d00100m_1.0m.tif"
#> [1] "Process: 18700 - 260966.8/6652718.0 - dom - 100/1 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dom_25833_d00100m_1.0m.tif"
#> [1] "Process: 18700 - 260966.8/6652718.0 - dtm - 20000/20 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dtm_25833_d20000m_20.0m.tif"
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#>    azimuth horizon_height       range
#> 1        0       9.939936   54.147945
#> 2      350       3.635535 3361.631301
#> 3      340       5.232884 3233.716566
#> 4      330      20.886836   30.413813
#> 5      320      12.180172   20.000000
#> 6      310      32.785109   36.069378
#> 7      300       6.553074   41.048752
#> 8      290       9.113168   17.464249
#> 9      280      10.758612   51.419841
#> 10     270       9.838308   55.145263
#> 11     260       7.956049   49.254441
#> 12     250       6.645440   69.641941
#> 13     240      12.569229   65.741920
#> 14     230      14.440466   80.212219
#> 15     220      11.846928   50.209561
#> 16     210      22.669420   38.275318
#> 17     200      63.481035    1.000000
#> 18     190      63.481035    1.000000
#> 19     180      63.481035    1.000000
#> 20     170      63.481035    1.000000
#> 21     160      63.481035    1.000000
#> 22     150      63.481035    1.000000
#> 23     140       7.332311   15.811388
#> 24     130       7.923324    9.219544
#> 25     120      12.459800   15.811388
#> 26     110       9.732914   53.712196
#> 27     100      10.646912   58.694122
#> 28      90       9.109575   64.124878
#> 29      80       7.823404   75.425460
#> 30      70      21.361401   52.201533
#> 31      60      22.801016   62.169124
#> 32      50      18.286524   47.201695
#> 33      40      18.566975   46.690470
#> 34      30      18.431249   46.872167
#> 35      20      17.705236   48.332184
#> 36      10       9.871761   52.392748
compute_horizon_max(stn, step = 1, f_output_all = TRUE)
#> [1] "Process: 18700 - 260966.8/6652718.0 - dtm - 100/1 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dtm_25833_d00100m_1.0m.tif"
#> [1] "Process: 18700 - 260966.8/6652718.0 - dom - 100/1 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dom_25833_d00100m_1.0m.tif"
#> [1] "Process: 18700 - 260966.8/6652718.0 - dtm - 20000/20 - path: data/dem"
#> [1] "Load existing file: data/dem/18700_dtm_25833_d20000m_20.0m.tif"
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#> Over-riding projection check
#> Importing raster map <elev>...
#>    0%   3%   6%   9%  12%  15%  18%  21%  24%  27%  30%  33%  36%  39%  42%  45%  48%  51%  54%  57%  60%  63%  66%  69%  72%  75%  78%  81%  84%  87%  90%  93%  96%  99% 100%
#> SpatRaster read into GRASS using r.in.gdal from memory
#>     azimuth horizon_max horizon_dem horizon_dsm horizon_demkm   range_max
#> 1         0    9.939936    2.186188    9.939936      2.696101   54.147945
#> 2       359    9.818483    2.186188    9.818483      2.748717   54.083269
#> 3       358    8.980175    2.186188    8.980175      2.824931   53.037722
#> 4       357    8.932049    2.186188    8.932049      2.827416   54.009258
#> 5       356    3.649215    2.186188    3.649215      2.938305   54.000000
#> 6       355    3.028576    2.186188    2.103104      3.028576 2640.747747
#> 7       354    3.185106    2.186188    2.103104      3.185106 3303.050936
#> 8       353    3.370702    2.186188    2.103104      3.370702 3326.101791
#> 9       352    3.428152    2.186188    2.103104      3.428152 3308.798481
#> 10      351    3.506618    2.186188    2.103104      3.506618 2913.558722
#> 11      350    3.635535    2.186188    2.306929      3.635535 3361.631301
#> 12      349    3.812960    2.342462    2.362784      3.812960 3609.522324
#> 13      348    4.006219    2.369612    2.787974      4.006219 3617.658683
#> 14      347    4.248279    2.369612    2.660042      4.248279 3771.407052
#> 15      346    4.496740    2.369612    2.362784      4.496740 3660.091500
#> 16      345    4.742044    2.369612    2.362784      4.742044 3651.775835
#> 17      344    4.849007    2.369612    2.711264      4.849007 3605.474996
#> 18      343    4.992842    2.369612    4.128500      4.992842 3517.018403
#> 19      342    5.151684    2.369612    4.115418      5.151684 3531.490273
#> 20      341    5.124650    2.443135    3.348565      5.124650 3464.476888
#> 21      340    5.232884    2.524184    3.249575      5.232884 3233.716566
#> 22      339    5.241329    2.737505    4.248157      5.241329 3182.091492
#> 23      338   12.396627    2.871264   12.396627      5.228359   30.675723
#> 24      337   14.947356    3.066822   14.947356      4.933201   28.792360
#> 25      336   17.024463    3.275602   17.024463      4.691980   31.016125
#> 26      335   19.541723    3.407784   19.541723      4.454191   30.083218
#> 27      334   19.752768    3.542790   19.752768      4.308661   30.463092
#> 28      333   19.808159    3.909981   19.808159      4.319852   29.546573
#> 29      332   20.007029    3.909981   20.007029      4.329987   30.870698
#> 30      331   20.570215    4.064581   20.570215      4.421455   29.966648
#> 31      330   20.886836    4.196910   20.886836      4.483229   30.413813
#> 32      329   20.886836    4.325543   20.886836      4.454186   30.413813
#> 33      328   20.886836    4.416155   20.886836      4.308349   30.413813
#> 34      327   19.515686    4.426731   19.515686      4.275461   29.529646
#> 35      326   18.541882    4.506299   18.541882      4.247916   30.016662
#> 36      325   16.694693    4.541567   16.694693      4.336075   31.384710
#> 37      324   12.455313    4.598833   12.455313      4.459268   30.528675
#> 38      323    8.565076    4.674241    8.565076      4.494375   19.416488
#> 39      322    8.565076    4.813674    8.565076      4.608544   19.416488
#> 40      321   12.180172    4.813674   12.180172      4.723912   20.000000
#> 41      320   12.180172    4.956519   12.180172      4.720467   20.000000
#> 42      319   13.380166    5.061769   13.380166      4.679855   19.209373
#> 43      318   13.380166    5.152752   13.380166      4.637103   19.209373
#> 44      317   14.966405    5.152752   14.966405      4.566724   19.849433
#> 45      316   14.966405    5.189911   14.966405      4.469136   19.849433
#> 46      315   15.627119    5.279638   15.627119      4.412360   20.518285
#> 47      314   15.627119    5.315918   15.627119      4.293234   20.518285
#> 48      313   15.627119    5.315918   15.627119      4.172790   20.518285
#> 49      312   23.089759    5.402727   23.089759      3.982083   35.355339
#> 50      311   32.785109    5.447979   32.785109      3.933547   36.069378
#> 51      310   32.785109    5.447979   32.785109      3.915017   36.069378
#> 52      309   23.693429    5.459985   23.693429      3.872358   36.796739
#> 53      308   15.124667    5.515332   15.124667      3.931329   20.518285
#> 54      307   11.878021    5.515332   11.878021      3.836350   19.849433
#> 55      306   11.878021    5.554455   11.878021      3.580584   19.849433
#> 56      305    8.745321    5.605095    8.745321      3.495702   18.439089
#> 57      304    6.424614    5.618201    6.424614      3.329188   20.615528
#> 58      303    6.349063    5.660726    6.349063      3.057981   42.201896
#> 59      302    6.789614    5.707200    6.789614      2.783159   41.617304
#> 60      301    6.789614    5.707200    6.789614      2.612539   41.617304
#> 61      300    6.553074    5.756375    6.553074      2.421350   41.048752
#> 62      299    9.303251    5.773905    9.303251      2.294708   17.492856
#> 63      298    9.303251    5.781203    9.303251      1.962171   17.492856
#> 64      297   11.055700    5.849459   11.055700      1.814816   18.357560
#> 65      296   11.055700    5.893722   11.055700      1.790055   18.357560
#> 66      295   11.055700    5.893722   11.055700      1.748793   18.357560
#> 67      294   11.641041    5.893054   11.641041      1.758536   17.888544
#> 68      293   11.641041    5.893054   11.641041      1.715176   17.888544
#> 69      292   11.641041    5.881490   11.641041      1.823076   17.888544
#> 70      291   11.641041    5.881490   11.641041      1.811709   17.888544
#> 71      290    9.113168    5.853562    9.113168      1.782262   17.464249
#> 72      289    9.113168    5.853562    9.113168      1.869417   17.464249
#> 73      288    6.270351    5.853562    6.270351      1.838719   18.384776
#> 74      287    6.270351    5.766615    6.270351      1.762932   18.384776
#> 75      286    5.954507    5.766615    5.954507      1.667244   44.598206
#> 76      285    6.231283    5.714870    6.231283      1.579895   43.324358
#> 77      284    6.018215    5.714870    6.018215      1.477209   52.497619
#> 78      283    5.931825    5.630658    5.931825      1.330745   46.840154
#> 79      282    8.284994    5.630658    8.284994      1.330745   51.923020
#> 80      281   10.226637    5.548126   10.226637      1.330745   51.662365
#> 81      280   10.758612    5.548126   10.758612      1.330745   51.419841
#> 82      279   11.298909    5.473187   11.298909      1.330745   51.195703
#> 83      278   11.363601    5.429069   11.363601      1.330745   50.990195
#> 84      277   10.939793    5.383023   10.939793      1.105357   51.971146
#> 85      276   10.341410    5.308883   10.341410      0.972902   50.803543
#> 86      275    8.309175    5.254644    8.309175      0.984322   50.635956
#> 87      274    8.182928    5.185130    8.182928      1.088563   77.646635
#> 88      273    8.367777    5.158213    8.367777      1.097234   77.524190
#> 89      272    8.359009    5.085464    8.359009      1.092197   77.414469
#> 90      271    9.128993    5.051297    9.128993      1.157858   54.147945
#> 91      270    9.838308    5.005610    9.838308      1.120984   55.145263
#> 92      269    9.232735    4.918894    9.232735      1.130129   55.081757
#> 93      268    8.693906    4.918894    8.693906      1.219334   49.010203
#> 94      267    9.083959    4.863863    9.083959      1.175226   55.009090
#> 95      266    7.158666    4.863863    7.158666      1.073177   50.000000
#> 96      265    6.461135    4.785644    6.461135      1.113241   49.010203
#> 97      264    7.432562    4.785644    7.432562      1.254863   49.040799
#> 98      263    7.333528    4.703331    7.333528      1.383239   47.095647
#> 99      262    7.926363    4.703331    7.926363      1.333968   49.162994
#> 100     261    7.956049    4.617900    7.956049      1.218266   49.254441
#> 101     260    7.956049    4.617900    7.956049      1.182818   49.254441
#> 102     259    7.296334    4.531595    7.296334      1.162403   74.545288
#> 103     258    7.619863    4.531595    7.619863      1.146093   75.802375
#> 104     257    7.342130    4.487359    7.342130      1.146487   79.075913
#> 105     256   10.536681    4.835981   10.536681      1.106551   77.278716
#> 106     255   11.308860    4.835981   11.308860      0.986326   77.466122
#> 107     254   11.227512    5.073095   11.227512      0.916134   78.854296
#> 108     253   10.228151    5.073095   10.228151      0.916134   77.129761
#> 109     252    9.513746    5.073095    9.513746      0.916134   78.339007
#> 110     251    8.986791    4.797520    8.986791      0.631089   49.729267
#> 111     250    6.645440    4.797520    6.645440      0.688372   69.641941
#> 112     249    6.219118    4.375778    6.219118      0.659758   37.643060
#> 113     248    5.863048    3.934260    5.863048      0.696854   64.195015
#> 114     247    5.598576    3.810657    5.598576      0.684058   64.513564
#> 115     246    6.440794    3.722110    6.440794      0.746024   65.192024
#> 116     245    7.811516    3.658710    7.811516      0.810180   63.324561
#> 117     244    9.168214    3.570034    9.168214      0.850557   62.769419
#> 118     243    9.346933    3.504126    9.346933      0.866220   63.158531
#> 119     242   10.963551    3.446847   10.963551      0.770822   64.884513
#> 120     241   12.133326    3.433947   12.133326      0.719827   65.306967
#> 121     240   12.569229    3.366296   12.569229      0.721263   65.741920
#> 122     239   12.594533    3.321164   12.594533      0.284792   66.189123
#> 123     238   12.186158    3.258746   12.186158      0.355331   64.884513
#> 124     237   12.116923    3.019199   12.116923      0.223650   65.368188
#> 125     236   11.468926    2.868602   11.468926      0.193486   64.498062
#> 126     235   11.582319    2.868602   11.582319      0.223100   66.887966
#> 127     234   12.867229    2.609607   12.867229      0.201080   67.416615
#> 128     233   12.935784    2.579390   12.935784      0.277278   67.955868
#> 129     232   12.286418    2.504146   12.286418      0.360941   79.881162
#> 130     231   14.403619    2.454151   14.403619      0.297915   79.630396
#> 131     230   14.440466    2.320107   14.440466      0.508417   80.212219
#> 132     229   14.774570    2.209069   14.774570      0.476171   45.607017
#> 133     228   15.460929    2.108450   15.460929      0.355905   47.010637
#> 134     227   15.347007    2.027161   15.347007      0.477458   46.227697
#> 135     226   15.725672    1.792968   15.725672      0.426781   48.918299
#> 136     225   15.761065    1.625543   15.761065      0.378091   49.578221
#> 137     224   15.761065    1.484396   15.761065      0.348247   49.578221
#> 138     223   15.369207    1.356685   15.369207      0.251930   48.104054
#> 139     222   14.680277    1.356685   14.680277      0.163510   48.795492
#> 140     221   12.558885    1.356685   12.558885      0.160775   49.497475
#> 141     220   11.846928    1.356685   11.846928      0.117412   50.209561
#> 142     219   12.595025    1.356685   12.595025      0.081239   33.970576
#> 143     218   16.123612    1.356685   16.123612      0.049060   33.301652
#> 144     217   17.802674    1.356685   17.802674      0.022953   34.058773
#> 145     216   20.115683    1.356685   20.115683      0.017667   33.421550
#> 146     215   20.115683    1.356685   20.115683      0.000000   33.421550
#> 147     214   20.987719    1.356685   20.987719      0.047309   32.802439
#> 148     213   21.781483    1.356685   21.781483      0.070996   33.600595
#> 149     212   21.781483    1.356685   21.781483      0.000000   33.600595
#> 150     211   21.040083    1.356685   21.040083      0.000000   33.015148
#> 151     210   22.669420    1.356685   22.669420      0.000000   38.275318
#> 152     209   23.127839    1.356685   23.127839      0.000000   37.735925
#> 153     208   23.433103    1.356685   23.433103      0.000000   38.587563
#> 154     207   24.133368    1.356685   24.133368      0.000000   38.078866
#> 155     206   63.481035   63.481035   63.301882      0.000000    1.000000
#> 156     205   63.481035   63.481035   63.301882      0.000000    1.000000
#> 157     204   63.481035   63.481035   63.301882      0.000000    1.000000
#> 158     203   63.481035   63.481035   63.301882      0.000000    1.000000
#> 159     202   63.481035   63.481035   63.301882      0.000000    1.000000
#> 160     201   63.481035   63.481035   63.301882      0.097973    1.000000
#> 161     200   63.481035   63.481035   63.301882      0.207243    1.000000
#> 162     199   63.481035   63.481035   63.301882      0.315284    1.000000
#> 163     198   63.481035   63.481035   63.301882      0.359731    1.000000
#> 164     197   63.481035   63.481035   63.301882      0.363428    1.000000
#> 165     196   63.481035   63.481035   63.301882      0.417247    1.000000
#> 166     195   63.481035   63.481035   63.301882      0.336594    1.000000
#> 167     194   63.481035   63.481035   63.301882      0.242955    1.000000
#> 168     193   63.481035   63.481035   63.301882      0.215107    1.000000
#> 169     192   63.481035   63.481035   63.301882      0.250452    1.000000
#> 170     191   63.481035   63.481035   63.301882      0.185893    1.000000
#> 171     190   63.481035   63.481035   63.301882      0.153583    1.000000
#> 172     189   63.481035   63.481035   63.301882      0.179474    1.000000
#> 173     188   63.481035   63.481035   63.301882      0.131751    1.000000
#> 174     187   63.481035   63.481035   63.301882      0.045321    1.000000
#> 175     186   63.481035   63.481035   63.301882      0.000000    1.000000
#> 176     185   63.481035   63.481035   63.301882      0.000000    1.000000
#> 177     184   63.481035   63.481035   63.301882      0.000000    1.000000
#> 178     183   63.481035   63.481035   63.301882      0.000000    1.000000
#> 179     182   63.481035   63.481035   63.301882      0.000000    1.000000
#> 180     181   63.481035   63.481035   63.301882      0.000000    1.000000
#> 181     180   63.481035   63.481035   63.301882      0.000000    1.000000
#> 182     179   63.481035   63.481035   63.301882      0.000000    1.000000
#> 183     178   63.481035   63.481035   63.301882      0.000000    1.000000
#> 184     177   63.481035   63.481035   63.301882      0.005470    1.000000
#> 185     176   63.481035   63.481035   63.301882      0.058974    1.000000
#> 186     175   63.481035   63.481035   63.301882      0.060290    1.000000
#> 187     174   63.481035   63.481035   63.301882      0.084605    1.000000
#> 188     173   63.481035   63.481035   63.301882      0.090393    1.000000
#> 189     172   63.481035   63.481035   63.301882      0.025195    1.000000
#> 190     171   63.481035   63.481035   63.301882      0.000000    1.000000
#> 191     170   63.481035   63.481035   63.301882      0.000000    1.000000
#> 192     169   63.481035   63.481035   63.301882      0.000000    1.000000
#> 193     168   63.481035   63.481035   63.301882      0.000000    1.000000
#> 194     167   63.481035   63.481035   63.301882      0.173273    1.000000
#> 195     166   63.481035   63.481035   63.301882      0.103432    1.000000
#> 196     165   63.481035   63.481035   63.301882      0.116651    1.000000
#> 197     164   63.481035   63.481035   63.301882      0.044141    1.000000
#> 198     163   63.481035   63.481035   63.301882      0.160110    1.000000
#> 199     162   63.481035   63.481035   63.301882      0.172970    1.000000
#> 200     161   63.481035   63.481035   63.301882      0.180253    1.000000
#> 201     160   63.481035   63.481035   63.301882      0.202930    1.000000
#> 202     159   63.481035   63.481035   63.301882      0.221064    1.000000
#> 203     158   63.481035   63.481035   63.301882      0.249543    1.000000
#> 204     157   63.481035   63.481035   63.301882      0.401851    1.000000
#> 205     156   63.481035   63.481035   63.301882      0.454624    1.000000
#> 206     155   63.481035   63.481035   63.301882      0.365863    1.000000
#> 207     154   63.481035   63.481035   63.301882      0.342837    1.000000
#> 208     153   63.481035   63.481035   63.301882      0.349755    1.000000
#> 209     152   63.481035   63.481035   63.301882      0.408427    1.000000
#> 210     151   63.481035   63.481035   63.301882      0.414663    1.000000
#> 211     150   63.481035   63.481035   63.301882      0.419475    1.000000
#> 212     149   63.481035   63.481035   63.301882      0.475064    1.000000
#> 213     148   63.481035   63.481035   63.301882      0.537608    1.000000
#> 214     147   63.481035   63.481035   63.301882      0.713490    1.000000
#> 215     146   10.646834    0.000000   10.646834      0.794909   99.478641
#> 216     145   10.047831    0.000000   10.047831      0.779382   15.264338
#> 217     144   10.047831    0.000000   10.047831      0.789860   15.264338
#> 218     143   10.047831    0.000000   10.047831      0.861901   15.264338
#> 219     142    9.352945    0.000000    9.352945      0.794279  118.713942
#> 220     141    8.736032    0.000000    8.736032      0.711481   18.027756
#> 221     140    7.332311    0.000000    7.332311      0.667301   15.811388
#> 222     139    7.490012    0.000000    7.490012      0.715246  126.210142
#> 223     138   11.092774    0.000000   11.092774      0.776849    9.219544
#> 224     137   11.092774    0.000000   11.092774      0.765983    9.219544
#> 225     136   11.092774    0.000000   11.092774      0.812703    9.219544
#> 226     135   11.092774    0.000000   11.092774      0.720187    9.219544
#> 227     134   11.092774    0.000000   11.092774      0.644220    9.219544
#> 228     133   11.092774    0.000000   11.092774      0.615248    9.219544
#> 229     132    7.365093    0.000000    7.365093      0.674670  106.075445
#> 230     131    7.437130    0.000000    7.437130      0.625159  107.480231
#> 231     130    7.923324    0.000000    7.923324      0.643772    9.219544
#> 232     129    7.923324    0.000000    7.923324      0.700064    9.219544
#> 233     128    8.881715    0.000000    8.881715      0.726784   15.620499
#> 234     127   10.079754    0.000000   10.079754      0.851565   14.212670
#> 235     126   10.079754    0.000000   10.079754      0.830951   14.212670
#> 236     125   12.578786    0.000000   12.578786      0.864529   15.000000
#> 237     124   12.578786    0.000000   12.578786      0.846182   15.000000
#> 238     123   12.578786    0.000000   12.578786      0.864455   15.000000
#> 239     122   12.578786    0.000000   12.578786      0.925121   15.000000
#> 240     121   12.578786    0.000000   12.578786      0.927857   15.000000
#> 241     120   12.459800    0.000000   12.459800      1.023147   15.811388
#> 242     119   12.459800    0.000000   12.459800      1.110735   15.811388
#> 243     118   12.442900    0.000000   12.442900      1.177178   14.422205
#> 244     117   11.354036    0.000000   11.354036      1.110374   13.892444
#> 245     116   11.354036    0.000000   11.354036      1.188912   13.892444
#> 246     115   11.354036    0.000000   11.354036      1.253615   13.892444
#> 247     114   11.354036    0.000000   11.354036      1.324318   13.892444
#> 248     113   10.405815    0.000000   10.405815      1.324888   14.764823
#> 249     112   11.354405    0.000000   11.354405      1.314729   54.129474
#> 250     111   11.354405    0.000000   11.354405      1.326233   54.129474
#> 251     110    9.732914    0.000000    9.732914      1.338146   53.712196
#> 252     109    9.420969    0.000000    9.420969      1.353169   55.154329
#> 253     108   10.792453    0.000000   10.792453      1.381845   57.008771
#> 254     107    9.885861    0.000000    9.885861      1.440511   56.648036
#> 255     106    9.169913    0.000000    9.169913      1.386136   60.745370
#> 256     105   10.955915    0.000000   10.955915      1.371611   55.973208
#> 257     104    9.646166    0.000000    9.646166      1.299390   55.659680
#> 258     103   10.152053    0.000000   10.152053      1.241556   57.280014
#> 259     102   10.896995    0.000000   10.896995      1.153361   58.240879
#> 260     101   10.872508    0.000000   10.872508      1.027595   57.974132
#> 261     100   10.646912    0.015580   10.646912      0.900336   58.694122
#> 262      99   10.822515    0.000000   10.822515      0.837562   58.463664
#> 263      98   11.070726    0.015981   11.070726      0.811939   58.249464
#> 264      97   10.762614    0.000000   10.762614      0.866234   58.051701
#> 265      96   10.265672    0.019722   10.265672      0.854202   58.855756
#> 266      95    9.275388    0.044353    9.275388      0.740705   62.801274
#> 267      94   10.731452    0.059983   10.731452      0.692149   63.639610
#> 268      93   10.353822    0.077928   10.353822      0.697983   64.498062
#> 269      92    9.220488    0.068631    9.220488      0.745542   62.289646
#> 270      91    9.148837    0.082804    9.148837      0.779932   65.192024
#> 271      90    9.109575    0.103166    9.109575      0.804863   64.124878
#> 272      89    8.779123    0.116283    8.779123      0.845073   67.067131
#> 273      88    8.821810    0.119134    8.821810      0.865316   73.027392
#> 274      87    9.798953    0.158782    9.798953      0.953257   70.007142
#> 275      86    9.642045    0.174956    9.642045      1.154299   69.000000
#> 276      85    9.018525    0.227990    9.018525      1.202281   73.027392
#> 277      84    9.523864    0.301920    9.523864      1.309109   73.061618
#> 278      83    9.032578    0.437725    9.032578      1.321977   72.111026
#> 279      82    8.316898    0.515840    8.316898      1.494826   72.173402
#> 280      81    8.699577    0.428449    8.699577      1.654872   75.325958
#> 281      80    7.823404    0.528814    7.823404      1.712764   75.425460
#> 282      79    7.445471    0.646708    7.445471      1.896336   75.663730
#> 283      78    6.967832    0.720991    6.967832      2.179312   79.906195
#> 284      77    7.320845    0.791772    7.320845      2.359957   79.075913
#> 285      76    6.539044    0.966451    6.539044      2.526465   85.328776
#> 286      75    6.002198    1.015499    6.002198      2.621462   54.129474
#> 287      74   16.484945    1.060354   16.484945      2.747695   54.341513
#> 288      73   21.780999    1.120053   21.780999      2.844147   53.366656
#> 289      72   21.482357    1.183380   21.482357      2.956572   52.630789
#> 290      71   22.018406    1.183516   22.018406      3.045772   51.923020
#> 291      70   21.361401    1.250808   21.361401      3.010026   52.201533
#> 292      69   22.161387    1.286466   22.161387      3.103118   51.244512
#> 293      68   23.750355    1.338733   23.750355      3.175855   57.870545
#> 294      67   23.506184    1.373317   23.506184      3.100555   58.189346
#> 295      66   24.801587    1.398688   24.801587      2.893572   56.648036
#> 296      65   25.946738    1.409497   25.946738      2.541609   54.781384
#> 297      64   27.437766    1.397541   27.437766      2.626473   55.154329
#> 298      63   26.739598    1.397541   26.739598      2.586862   56.462377
#> 299      62   25.316201    1.401753   25.316201      2.693726   56.859476
#> 300      61   23.501764    1.399312   23.501764      2.666167   57.271284
#> 301      60   22.801016    1.410301   22.801016      2.714658   62.169124
#> 302      59   21.511611    1.414417   21.511611      2.667229   63.953108
#> 303      58   18.779931    1.538243   18.779931      2.649748   62.177166
#> 304      57   18.300963    1.629143   18.300963      2.375607   64.007812
#> 305      56   17.787754    1.882557   17.787754      2.118467   48.877398
#> 306      55   17.787754    1.914549   17.787754      2.100374   48.877398
#> 307      54   17.926382    1.765678   17.926382      2.117075   48.548944
#> 308      53   18.069458    1.840291   18.069458      2.099162   47.707442
#> 309      52   17.984542    1.915564   17.984542      2.022242   48.259714
#> 310      51   18.142710    2.079687   18.142710      2.016178   48.010416
#> 311      50   18.286524    2.048866   18.286524      2.049505   47.201695
#> 312      49   18.215487    1.993499   18.215487      2.038042   47.801674
#> 313      48   18.370244    1.857969   18.370244      2.008877   47.010637
#> 314      47   18.502649    1.631028   18.502649      1.987545   46.227697
#> 315      46   18.558949    1.345984   18.558949      1.962073   46.097722
#> 316      45   18.558949    1.355589   18.558949      1.931920   46.097722
#> 317      44   18.412291    1.352112   18.412291      1.978786   47.423623
#> 318      43   18.536228    1.346526   18.536228      2.042906   46.690470
#> 319      42   18.432529    1.333693   18.432529      2.044117   47.381431
#> 320      41   18.434721    1.407660   18.434721      2.080004   47.381431
#> 321      40   18.566975    1.461008   18.566975      2.097598   46.690470
#> 322      39   18.566975    1.536002   18.566975      2.129448   46.690470
#> 323      38   18.567362    1.587141   18.567362      2.158333   46.754679
#> 324      37   18.672524    1.594230   18.672524      2.148760   46.097722
#> 325      36   18.531737    1.630068   18.531737      2.192344   46.861498
#> 326      35   18.494164    1.647414   18.494164      2.211186   47.010637
#> 327      34   18.583936    1.669678   18.583936      2.280142   46.400431
#> 328      33   18.583936    1.724072   18.583936      2.326779   46.400431
#> 329      32   18.505177    1.741025   18.505177      2.376151   46.615448
#> 330      31   18.505177    1.787218   18.505177      2.305462   46.615448
#> 331      30   18.431249    1.767167   18.431249      2.283979   46.872167
#> 332      29   18.489577    1.753852   18.489577      2.264404   46.324939
#> 333      28   18.368365    1.770624   18.368365      2.294256   46.647615
#> 334      27   18.368365    1.754254   18.368365      2.449459   46.647615
#> 335      26   18.282648    2.040976   18.282648      2.536586   47.010637
#> 336      25   18.282648    2.040976   18.282648      2.522086   47.010637
#> 337      24   17.953673    2.040976   17.953673      2.508580   48.764741
#> 338      23   17.815072    2.040976   17.815072      2.430304   49.193496
#> 339      22   17.838968    2.040976   17.838968      2.487892   48.754487
#> 340      21   17.705236    2.040976   17.705236      2.527755   48.332184
#> 341      20   17.705236    2.040976   17.705236      2.519017   48.332184
#> 342      19   17.692340    2.040976   17.692340      2.665440   48.846699
#> 343      18   17.515246    2.040976   17.515246      2.622832   49.769469
#> 344      17   17.171053    2.040976   17.171053      2.594957   50.328918
#> 345      16   15.861276    2.040976   15.861276      2.522477   49.979996
#> 346      15    9.893121    2.040976    9.893121      2.453487   51.865210
#> 347      14    9.745260    2.040976    9.745260      2.435294   51.546096
#> 348      13    9.610513    2.040976    9.610513      2.416636   52.201533
#> 349      12    9.577626    2.040976    9.577626      2.381419   38.327536
#> 350      11    9.666357    2.040976    9.666357      2.333582   52.630789
#> 351      10    9.871761    2.183938    9.871761      2.232334   52.392748
#> 352       9   10.064807    2.183938   10.064807      2.293253   52.172790
#> 353       8   10.268221    2.183938   10.268221      2.397573   50.990195
#> 354       7   10.597030    2.183938   10.597030      2.625105   48.836462
#> 355       6   10.637235    2.183938   10.637235      2.643227   48.662100
#> 356       5   10.596047    2.186188   10.596047      2.562959   49.648766
#> 357       4   10.460202    2.186188   10.460202      2.596136   50.487622
#> 358       3   10.326412    2.186188   10.326412      2.545676   50.358713
#> 359       2   10.131129    2.186188   10.131129      2.543737   52.239832
#> 360       1   10.014721    2.186188   10.014721      2.597729   52.153619
#>      range_dem  range_dsm range_demkm
#> 1     3.000000  54.147945  1523.32400
#> 2     3.000000  54.083269  1502.16937
#> 3     3.000000  53.037722  2140.89448
#> 4     3.000000  54.009258  2100.14774
#> 5     3.000000  54.000000  2680.14163
#> 6     3.000000   3.000000  2640.74775
#> 7     3.000000   3.000000  3303.05094
#> 8     3.000000   3.000000  3326.10179
#> 9     3.000000   3.000000  3308.79848
#> 10    3.000000   3.000000  2913.55872
#> 11    3.000000  68.468971  3361.63130
#> 12    4.123106   4.123106  3609.52232
#> 13    3.162278  56.568542  3617.65868
#> 14    3.162278  57.706152  3771.40705
#> 15    3.162278   4.123106  3660.09150
#> 16    3.162278   4.123106  3651.77583
#> 17    3.162278 101.414989  3605.47500
#> 18    3.162278 102.839681  3517.01840
#> 19    3.162278 103.077641  3531.49027
#> 20   36.400549 103.580886  3464.47689
#> 21   34.481879 104.120123  3233.71657
#> 22   35.735137  29.410882  3182.09149
#> 23   33.837849  30.675723  3118.09014
#> 24   35.114100  28.792360  3048.74806
#> 25   33.241540  31.016125  2961.29000
#> 26   34.539832  30.083218  2874.78924
#> 27   33.615473  30.463092  4869.37219
#> 28   31.780497  29.546573  4874.21664
#> 29   31.780497  30.870698  4878.03134
#> 30   32.202484  29.966648  4804.20170
#> 31   31.304952  30.413813  4821.11627
#> 32   31.764760  30.413813  4812.43784
#> 33   30.886890  30.413813  5566.07124
#> 34   30.016662  29.529646  5095.57634
#> 35   31.384710  30.016662  5279.26069
#> 36   30.528675  31.384710  4685.58403
#> 37   31.064449  30.528675  4738.43039
#> 38   30.232433  19.416488  4864.89294
#> 39   29.410882  19.416488  4876.39097
#> 40   29.410882  20.000000  4877.90833
#> 41   30.000000  20.000000  4836.40047
#> 42   29.206164  19.209373  4780.28687
#> 43   28.425341  19.209373  4841.60841
#> 44   28.425341  19.849433  4804.28496
#> 45   29.068884  19.849433  4980.92767
#> 46   28.319605  20.518285  4946.43541
#> 47   27.586228  20.518285  4969.83248
#> 48   27.586228  20.518285  4923.61778
#> 49   28.284271  35.355339  4865.34516
#> 50   27.586228  36.069378  3549.79296
#> 51   27.586228  36.069378  3550.46902
#> 52   28.319605  36.796739  3920.55717
#> 53   27.658633  20.518285  3981.30538
#> 54   27.658633  19.849433  3985.12158
#> 55   27.018512  19.849433  3905.22247
#> 56   27.802878  18.439089  3898.19956
#> 57   28.600699  20.615528  3848.62840
#> 58   27.202941  42.201896  3844.10442
#> 59   28.017851  41.617304  3696.79365
#> 60   28.017851  41.617304  3678.46116
#> 61   27.459060  41.048752  3705.49410
#> 62   28.301943  17.492856  3733.72428
#> 63   26.925824  17.492856  3691.05421
#> 64   27.784888  18.357560  3354.60320
#> 65   28.653098  18.357560 11949.80919
#> 66   28.653098  18.357560 11448.06327
#> 67   28.178006  17.888544 11764.13701
#> 68   28.178006  17.888544  1672.64458
#> 69   28.635642  17.888544  1681.94573
#> 70   28.635642  17.888544  1602.28848
#> 71   28.231188  17.464249  1620.53426
#> 72   28.231188  17.464249  9601.78158
#> 73   28.231188  18.384776  9704.71581
#> 74   28.792360  18.384776  9639.76380
#> 75   28.792360  44.598206  9837.42032
#> 76   28.460499  43.324358 10619.28615
#> 77   28.460499  52.497619  1619.17605
#> 78   28.160256  46.840154    44.72248
#> 79   28.160256  51.923020    44.72248
#> 80   28.861739  51.662365    44.72248
#> 81   28.861739  51.419841    44.72248
#> 82   29.832868  51.195703    44.72248
#> 83   28.635642  50.990195    44.72248
#> 84   29.614186  51.971146 10626.36560
#> 85   28.442925  50.803543 15746.69481
#> 86   30.413813  50.635956 10502.85272
#> 87   29.274562  77.646635 14411.53650
#> 88   30.265492  77.524190 14297.05634
#> 89   29.154759  77.414469 10653.13448
#> 90   30.149627  54.147945 10656.66355
#> 91   28.071338  55.145263 10281.46319
#> 92   29.017236  55.081757 10211.54315
#> 93   29.017236  49.010203  9504.36302
#> 94   29.000000  55.009090  9460.99759
#> 95   29.000000  50.000000  4160.15208
#> 96   29.017236  49.010203  4101.32186
#> 97   29.017236  49.040799  4123.20870
#> 98   30.066593  47.095647  4147.05436
#> 99   30.066593  49.162994 11291.57096
#> 100  30.149627  49.254441  4237.18039
#> 101  30.149627  49.254441  4083.87486
#> 102  30.265492  74.545288  3912.18165
#> 103  30.265492  75.802375 11642.14657
#> 104  30.413813  79.075913  3952.26277
#> 105  25.495098  77.278716  3982.10909
#> 106  25.495098  77.466122  3996.94868
#> 107  25.709920  78.854296    40.00100
#> 108  25.709920  77.129761    40.00100
#> 109  25.709920  78.339007    40.00100
#> 110  25.961510  49.729267 17476.69509
#> 111  25.961510  69.641941 20418.11069
#> 112  25.000000  37.643060 20458.64431
#> 113  27.202941  64.195015 20194.92983
#> 114  31.622777  64.513564 20209.46363
#> 115  32.893768  65.192024 21236.72642
#> 116  34.176015  63.324561 21226.42050
#> 117  35.468296  62.769419 21222.48145
#> 118  36.769553  63.158531 21224.91281
#> 119  42.544095  64.884513 20929.12565
#> 120  42.953463  65.306967 17974.40829
#> 121  41.146081  65.741920 17889.21465
#> 122  40.249224  66.189123 22470.06997
#> 123  40.706265  64.884513 22701.83622
#> 124  38.948684  65.368188 22932.83006
#> 125  39.446166  64.498062 23062.03356
#> 126  39.446166  66.887966 23341.78316
#> 127  52.153619  67.416615 23155.40558
#> 128  52.697249  67.955868 23847.81824
#> 129  51.865210  79.881162 23881.57690
#> 130  54.644304  79.630396 22413.36023
#> 131  56.612719  80.212219 22365.93416
#> 132  56.400355  45.607017 22580.99852
#> 133  57.008771  47.010637 22430.87061
#> 134  57.628118  46.227697 25470.87036
#> 135  57.489129  48.918299 25644.54876
#> 136  58.137767  49.578221 26163.98720
#> 137  61.619802  49.578221 26491.87592
#> 138   1.414214  48.104054 26572.25814
#> 139   1.414214  48.795492 26519.52418
#> 140   1.414214  49.497475 25173.88498
#> 141   1.414214  50.209561 24019.34315
#> 142   1.414214  33.970576 24116.87128
#> 143   1.414214  33.301652 23542.38117
#> 144   1.414214  34.058773 24051.76341
#> 145   1.414214  33.421550 24440.29184
#> 146   1.414214  33.421550     0.00000
#> 147   1.414214  32.802439 25147.68417
#> 148   1.414214  33.600595 24948.64968
#> 149   1.414214  33.600595     0.00000
#> 150   1.414214  33.015148     0.00000
#> 151   1.414214  38.275318     0.00000
#> 152   1.414214  37.735925     0.00000
#> 153   1.414214  38.587563     0.00000
#> 154   1.414214  38.078866     0.00000
#> 155   1.000000   1.000000     0.00000
#> 156   1.000000   1.000000     0.00000
#> 157   1.000000   1.000000     0.00000
#> 158   1.000000   1.000000     0.00000
#> 159   1.000000   1.000000     0.00000
#> 160   1.000000   1.000000 14730.26981
#> 161   1.000000   1.000000 13478.17060
#> 162   1.000000   1.000000 14163.47520
#> 163   1.000000   1.000000 12835.58481
#> 164   1.000000   1.000000 13324.37153
#> 165   1.000000   1.000000 13278.90116
#> 166   1.000000   1.000000 13326.35281
#> 167   1.000000   1.000000 13693.91748
#> 168   1.000000   1.000000 14078.49103
#> 169   1.000000   1.000000 14277.68161
#> 170   1.000000   1.000000 14044.31065
#> 171   1.000000   1.000000 17831.87871
#> 172   1.000000   1.000000 16298.45641
#> 173   1.000000   1.000000 16404.04384
#> 174   1.000000   1.000000 16937.81402
#> 175   1.000000   1.000000     0.00000
#> 176   1.000000   1.000000     0.00000
#> 177   1.000000   1.000000     0.00000
#> 178   1.000000   1.000000     0.00000
#> 179   1.000000   1.000000     0.00000
#> 180   1.000000   1.000000     0.00000
#> 181   1.000000   1.000000     0.00000
#> 182   1.000000   1.000000     0.00000
#> 183   1.000000   1.000000     0.00000
#> 184   1.000000   1.000000 19941.94282
#> 185   1.000000   1.000000 18620.73403
#> 186   1.000000   1.000000 19485.45569
#> 187   1.000000   1.000000 19396.17514
#> 188   1.000000   1.000000 19392.85444
#> 189   1.000000   1.000000 19857.23331
#> 190   1.000000   1.000000     0.00000
#> 191   1.000000   1.000000     0.00000
#> 192   1.000000   1.000000     0.00000
#> 193   1.000000   1.000000     0.00000
#> 194   1.000000   1.000000 18465.79131
#> 195   1.000000   1.000000 18030.31486
#> 196   1.000000   1.000000 16195.67580
#> 197   1.000000   1.000000 17198.75546
#> 198   1.000000   1.000000 16998.60133
#> 199   1.000000   1.000000 16696.72838
#> 200   1.000000   1.000000 14556.02259
#> 201   1.000000   1.000000 16316.41231
#> 202   1.000000   1.000000 15757.66527
#> 203   1.000000   1.000000 15742.22317
#> 204   1.000000   1.000000 14966.20189
#> 205   1.000000   1.000000 14459.75290
#> 206   1.000000   1.000000 14362.55944
#> 207   1.000000   1.000000 17192.91654
#> 208   1.000000   1.000000  5796.04000
#> 209   1.000000   1.000000 16989.55066
#> 210   1.000000   1.000000  5770.75948
#> 211   1.000000   1.000000 16153.91739
#> 212   1.000000   1.000000  7907.78898
#> 213   1.000000   1.000000  7776.80036
#> 214   1.000000   1.000000  7325.78209
#> 215   0.000000  99.478641  7350.96604
#> 216   0.000000  15.264338  7090.02915
#> 217   0.000000  15.264338  6941.21091
#> 218   0.000000  15.264338  6945.12863
#> 219   0.000000 118.713942  6823.30769
#> 220   0.000000  18.027756  6798.34618
#> 221   0.000000  15.811388 12244.78242
#> 222   0.000000 126.210142 12100.71573
#> 223   0.000000   9.219544 12232.08408
#> 224   0.000000   9.219544 12320.32423
#> 225   0.000000   9.219544 12174.57230
#> 226   0.000000   9.219544 13108.38629
#> 227   0.000000   9.219544 12826.85566
#> 228   0.000000   9.219544 13341.83270
#> 229   0.000000 106.075445 12969.59561
#> 230   0.000000 107.480231 13181.07305
#> 231   0.000000   9.219544 12873.42557
#> 232   0.000000   9.219544 15952.43942
#> 233   0.000000  15.620499 15767.41038
#> 234   0.000000  14.212670 12198.30479
#> 235   0.000000  14.212670 11846.30814
#> 236   0.000000  15.000000 15652.12602
#> 237   0.000000  15.000000 15570.00067
#> 238   0.000000  15.000000 14108.46385
#> 239   0.000000  15.000000 11150.77651
#> 240   0.000000  15.000000 10624.25732
#> 241   0.000000  15.811388 11094.29376
#> 242   0.000000  15.811388 10946.43280
#> 243   0.000000  14.422205 10873.98326
#> 244   0.000000  13.892444 10276.59867
#> 245   0.000000  13.892444 10865.72147
#> 246   0.000000  13.892444 10847.87000
#> 247   0.000000  13.892444 10744.14127
#> 248   0.000000  14.764823 10626.06444
#> 249   0.000000  54.129474 10448.55293
#> 250   0.000000  54.129474 10327.25193
#> 251   0.000000  53.712196 10200.52951
#> 252   0.000000  55.154329 10147.96279
#> 253   0.000000  57.008771 10521.80285
#> 254   0.000000  56.648036 10089.21649
#> 255   0.000000  60.745370 10387.67515
#> 256   0.000000  55.973208 10473.84766
#> 257   0.000000  55.659680 10139.38559
#> 258   0.000000  57.280014 10149.38177
#> 259   0.000000  58.240879 10239.25983
#> 260   0.000000  57.974132 10400.89461
#> 261  65.969690  58.694122  9716.39440
#> 262   0.000000  58.463664  9553.89783
#> 263  70.405966  58.249464 18094.48452
#> 264   0.000000  58.051701 10034.15338
#> 265  72.993150  58.855756  9923.49352
#> 266  71.847060  62.801274 10075.56826
#> 267  72.691127  63.639610 10030.16600
#> 268  73.552702  64.498062  4368.81612
#> 269  74.330344  62.289646  4682.76589
#> 270  66.189123  65.192024  4956.29178
#> 271  66.121101  64.124878  4729.64253
#> 272  67.067131  67.067131  4745.22137
#> 273  68.029405  73.027392  4742.18561
#> 274  69.007246  70.007142  6260.66767
#> 275  57.000000  69.000000  6400.28500
#> 276  57.008771  73.027392  6301.71290
#> 277  57.035077  73.061618  6365.47138
#> 278  57.078893  72.111026  6370.33979
#> 279  57.140179  72.173402  4813.60141
#> 280  60.299254  75.325958  4881.84305
#> 281  60.406953  75.425460  4989.43330
#> 282  61.522354  75.663730  4981.40951
#> 283  61.660360  79.906195  4831.10411
#> 284  62.801274  79.075913  4843.34315
#> 285  65.115282  85.328776  4899.79595
#> 286  65.306967  54.129474  4710.82897
#> 287  68.658576  54.341513  4746.65424
#> 288  68.883960  53.366656  5016.09988
#> 289  71.309186  52.630789  5078.90635
#> 290  70.342022  51.923020  5123.91576
#> 291  62.641839  52.201533  5000.12500
#> 292  62.936476  51.244512  4984.90145
#> 293  65.459911  57.870545  5015.50174
#> 294  66.730802  58.189346  5060.16603
#> 295  67.067131  56.648036  5049.40341
#> 296  67.416615  54.781384  5578.70559
#> 297  67.230945  55.154329  5470.96852
#> 298  67.230945  56.462377  5402.09765
#> 299  67.119297  56.859476  5550.09370
#> 300  66.211781  57.271284  5413.08277
#> 301  67.082039  62.169124  5536.67162
#> 302  66.189123  63.953108  5492.38639
#> 303  94.413982  62.177166  5467.20170
#> 304  95.383437  64.007812  5452.95211
#> 305  98.615415  48.877398  5701.08981
#> 306  99.126182  48.877398  5779.72925
#> 307  96.566040  48.548944  5681.83193
#> 308  99.569072  47.707442  5735.96069
#> 309  95.692215  48.259714  8853.62746
#> 310  91.021975  48.010416  8884.02759
#> 311  93.005376  47.201695  2560.14213
#> 312  96.400207  47.801674  2640.14176
#> 313 100.424101  47.010637  2692.64972
#> 314 101.671038  46.227697  2701.91876
#> 315  58.898217  46.097722  2824.60597
#> 316  59.548300  46.097722  2863.63580
#> 317  59.464275  47.423623  2704.21257
#> 318  58.728187  46.690470  2730.41622
#> 319  58.000000  47.381431  2743.78870
#> 320  43.139309  47.381431  2870.96014
#> 321  42.449971  46.690470  3013.18004
#> 322  43.185646  46.690470  2986.77854
#> 323  42.520583  46.754679  2974.76763
#> 324  41.868843  46.097722  2906.96104
#> 325  42.638011  46.861498  3180.51973
#> 326  43.416587  47.010637  3186.04889
#> 327  42.801869  46.400431  3052.28862
#> 328  43.600459  46.400431  3116.10101
#> 329  43.011626  46.615448  3180.64550
#> 330  43.829214  46.615448  3206.63595
#> 331  43.266615  46.872167  3272.81768
#> 332  44.102154  46.324939  2679.84310
#> 333  43.566042  46.647615  2675.58553
#> 334  43.908997  46.647615  2949.92120
#> 335   1.000000  47.010637  2964.53022
#> 336   1.000000  47.010637  3106.26503
#> 337   1.000000  48.764741  3006.00263
#> 338   1.000000  49.193496  2441.86366
#> 339   1.000000  48.754487  2799.85570
#> 340   1.000000  48.332184  2729.53705
#> 341   1.000000  48.332184  2731.36834
#> 342   1.000000  48.846699  2232.36403
#> 343   1.000000  49.769469  2198.69091
#> 344   1.000000  50.328918  2139.58615
#> 345   1.000000  49.979996  2106.89865
#> 346   1.000000  51.865210  2156.81074
#> 347   1.000000  51.546096  2163.19994
#> 348   1.000000  52.201533  2170.49201
#> 349   1.000000  38.327536  2140.14696
#> 350   1.000000  52.630789  2129.65414
#> 351   2.000000  52.392748  2100.43343
#> 352   2.000000  52.172790  1679.08932
#> 353   2.000000  50.990195  1694.50393
#> 354   2.000000  48.836462  1667.25492
#> 355   2.000000  48.662100  1663.77234
#> 356   3.000000  49.648766  1660.52337
#> 357   3.000000  50.487622  1250.31123
#> 358   3.000000  50.358713  1570.38954
#> 359   3.000000  52.239832  1568.22287
#> 360   3.000000  52.153619  1544.70687