Get station and sensor metadata from Frost API

Fetch station metadata based on station number & parameter id from Frost v1. The function gets station name, location and sensor details such as level, exposure and performance

get_metadata_frost(
  stationid = 18700,
  paramid = NULL,
  dx = 100,
  resx = 1,
  path = sprintf("output/%i", stationid)
)

Arguments

stationid

A station number as integer and defined by met.no

paramid

A parameter number as integer and defined by met.no

dx

A distance in metre or radius defining the extent of the bounding box from the centre point

resx

A horizontal resolution in metre

path

A directory path defining where will be saved project plots/data

Value

Station metadata

References

https://frost-beta.met.no/docs/codeexamples

Examples

get_metadata_frost(stationid = 18700)
#> [1] " "
#> [1] "-------------------------------------------"
#> [1] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [2] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [3] "station MET.NO: 18700 -- OSLO - BLINDERN -- WMO: 0-20000-0-01492"      
#> [4] "station MET.NO: 18700 -- OSLO - BLINDERN -- WIGOS: 0-20000-0-01492"    
#> [1] "-------------------------------------------"
#> [1] " "
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 1, 26  (geometries, attributes)
#>  extent      : 260966.8, 260966.8, 6652718, 6652718  (xmin, xmax, ymin, ymax)
#>  coord. ref. : ETRS89 / UTM zone 33N (EPSG:25833) 
#>  names       : level parameterid sensor stationid  elev   lat   lon elev.1
#>  type        : <int>       <int>  <int>     <int> <num> <num> <num>  <num>
#>  values      :     0        3019      0     18700    94 59.94 10.72     94
#>     element.name    station.name (and 16 more)
#>            <chr>           <chr>              
#>  Minimum air te~ OSLO - BLINDERN              
get_metadata_frost(stationid = 18700, paramid = 211)
#> [1] " "
#> [1] "-------------------------------------------"
#> [1] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [2] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [3] "station MET.NO: 18700 -- OSLO - BLINDERN -- WMO: 0-20000-0-01492"      
#> [4] "station MET.NO: 18700 -- OSLO - BLINDERN -- WIGOS: 0-20000-0-01492"    
#> [1] "-------------------------------------------"
#> [1] " "
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 1, 26  (geometries, attributes)
#>  extent      : 260966.8, 260966.8, 6652718, 6652718  (xmin, xmax, ymin, ymax)
#>  coord. ref. : ETRS89 / UTM zone 33N (EPSG:25833) 
#>  names       : level parameterid sensor stationid  elev   lat   lon elev.1
#>  type        : <int>       <int>  <int>     <int> <num> <num> <num>  <num>
#>  values      :     0         211      0     18700    94 59.94 10.72     94
#>     element.name    station.name (and 16 more)
#>            <chr>           <chr>              
#>  Air temperature OSLO - BLINDERN              
get_metadata_frost(stationid = 18700, paramid = 211, dx = 100, resx = 1)
#> [1] " "
#> [1] "-------------------------------------------"
#> [1] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [2] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.11:6785"
#> [3] "station MET.NO: 18700 -- OSLO - BLINDERN -- WMO: 0-20000-0-01492"      
#> [4] "station MET.NO: 18700 -- OSLO - BLINDERN -- WIGOS: 0-20000-0-01492"    
#> [1] "-------------------------------------------"
#> [1] " "
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 1, 26  (geometries, attributes)
#>  extent      : 260966.8, 260966.8, 6652718, 6652718  (xmin, xmax, ymin, ymax)
#>  coord. ref. : ETRS89 / UTM zone 33N (EPSG:25833) 
#>  names       : level parameterid sensor stationid  elev   lat   lon elev.1
#>  type        : <int>       <int>  <int>     <int> <num> <num> <num>  <num>
#>  values      :     0         211      0     18700    94 59.94 10.72     94
#>     element.name    station.name (and 16 more)
#>            <chr>           <chr>              
#>  Air temperature OSLO - BLINDERN