Plot station with Ortophoto, grid and buffers

Overlay a grid and two buffers to count the area of land cover types and assess heat source proximity to a weather station and their distribution. Four different scales are plotted: 10 m, 30 m, 100 m and 1000 m.

Usage

plot_station_grid(
  stn = NULL,
  tile_name = "ortofoto",
  path = stn$path,
  grid_scale = c(10, 50, 100, 1000)
)

Arguments

stn

A SpatVector with station attributes from "get_latlon_frost"

tile_name

A string naming a tile to pass to "get_tile_wms"

path

A directory path defining where will be saved the plots, if path is NULL the plots are printed to the console

grid_scale

A value or array setting plot extent and grid parameters that would be looped through if `grid_scale` is an array

Value

A ggplot2 object

Examples

# Get station metadata
stn    <- get_metadata_frost(stationid = 18700)
#> [1] " "
#> [1] "-------------------------------------------"
#> [1] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.15" 
#> [2] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.15" 
#> [3] "station MET.NO: 18700 -- OSLO - BLINDERN -- WMO: 0-20000-0-01492"  
#> [4] "station MET.NO: 18700 -- OSLO - BLINDERN -- Fast_IP: 10.240.10.15" 
#> [5] "station MET.NO: 18700 -- OSLO - BLINDERN -- WIGOS: 0-20000-0-01492"
#> [1] "-------------------------------------------"
#> [1] " "

# Plot grid and buffers on four different scales
#plot_station_grid(stn)