Weather station context

Plots weather station's sun diagram and background maps to assess the exposure of a station and to compute its WMO/met.no siting classification

plot_station_siting_context(
  stationid = 18700,
  paramid = 211,
  f_verbose = FALSE,
  f_pdf = FALSE
)

Arguments

stationid

A station number as integer and defined by met.no

paramid

A parameter number as integer and defined by met.no

f_verbose

A boolean string to print debug messages, default is TRUE

f_pdf

A boolean string to combine all plots into a pdf file

Value

None

References

https://community.wmo.int/en/activity-areas/imop/siting-classification#MembersTools

Examples

# Plot sun diagram and map infos for a weather station
#plot_station_siting_context(stationid = 18700)
#plot_station_siting_context(stationid = 18700,
#                            paramid = 211,
#                            f_verbose = TRUE)