Get station metadata/coordinates from Frost API

Fetch stations metadata based on country, parameter id, ... using Frost v1. To handle such a large request output, a pagination protocol is used. The function gets station names, locations and alternate ids.

Usage

get_metadata_pagination(
  country = "norge",
  paramid = 211,
  timestamp = "2000-01-01T00:00:00Z/2024-01-01T00:00:00Z",
  hdrshow =
    "{\"id\":{},\"extra\":{\"station\":{\"location\":[],\"alternateids\":[]}}}",
  path = sprintf("output/stn_%s_%i", country, paramid)
)

Arguments

country

A country name as string and defined by met.no

paramid

A parameter number as integer and defined by met.no

timestamp

A time range as timestamp or string "latest"

hdrshow

A JSON string containing metadata structure to extract

path

A path where will be saved the metadata as SpatVector `.gpkg`

Value

Stations metadata

References

https://frost-beta.met.no/docs/codeexamples

Examples

get_metadata_pagination(country = "norge", paramid = 211)
#> Warning: GDAL Message 6: dataset output/stn_norge_211/stn_norge_211.shp does not support layer creation option OVERWRITE
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 657, 8  (geometries, attributes)
#>  extent      : -770349.4, 653347, -684164.7, 1286239  (xmin, xmax, ymin, ymax)
#>  coord. ref. : +proj=lcc +lat_0=63 +lon_0=15 +lat_1=63 +lat_2=63 +x_0=0 +y_0=0 +R=6371000 +units=m +no_defs 
#>  names       : level parameterid sensor stationid           WIGOS  elev   lat
#>  type        : <int>       <int>  <int>     <int>           <chr> <num> <num>
#>  values      :     0         211      0       180              NA   360 61.29
#>                    0         211      0       210     0-578-0-210  1020 61.32
#>                    0         211      0       700 0-20000-0-01393   672 61.89
#>    lon
#>  <num>
#>  12.27
#>  12.19
#>  12.05